{{cookiecutter.repo_name}}/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
nose
Flask-Testing
factory-boy>=2.2.1

# Management script
Flask-Script
